Namespace the image attrs with r:
The images were being found, but they couldn't be mapped to the right
relationship because the attr value is namespaced. I'm just adding
the 'r:' namespace to the values.

Note that I'm not familiar enough with the file format or this gem
to say that this is the right way for all docx files. I have only tested
with a collection of Word 2011 files all from the same sender.